+++ sign/policy.go
@@ -35,16 +35,16 @@ func (p CompletePolicy) Check(m ParticipationMask) bool {
 // least the given threshold number of participants t have cosigned to make a
 // collective signature valid.
 type ThresholdPolicy struct {
-	thold int
+	thold int64
 }
 
 // NewThresholdPolicy returns a new ThresholdPolicy with the given threshold.
 func NewThresholdPolicy(thold int) *ThresholdPolicy {
-	return &ThresholdPolicy{thold: thold}
+	return &ThresholdPolicy{thold: int64(thold)}
 }
 
 // Check verifies that at least a threshold number of participants have
 // contributed to a collective signature.
 func (p ThresholdPolicy) Check(m ParticipationMask) bool {
-	return m.CountEnabled() >= p.thold
+	return int64(m.CountEnabled()) >= p.thold
 }
